split current staff routes
